+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New
 
 * CallableMethodsList value builder
+* FirstMethodMatchingType value builder
 
 ## 2.1.1 - Sat Jun 27 2015